Added Wave, EnemyFactory, and LevelManager
Wave is a class that gives all information on specific enemies, the count, and if the wave is cleared.
EnemyFactory produces enemies and can either produce a single enemy, a wave from a list of enemies,
or a wave from a json.
LevelManager manages the waves and the enemies, how they spawn, and there is also unimplemented traits
for a type of progressive level system with Record<number, Wave[]<